refactor(TextEditor): replace hook with normal function
