+++ packages/picasso-lab/src/TextEditor/hooks/getTextChangeHandler/index.ts
@@ -1,0 +1,1 @@
+export { default } from './getTextChangeHandler'
