--- packages/picasso-lab/src/TextEditor/hooks/useTextChangeHandler/index.ts
@@ -1,1 +1,0 @@
-export { default } from './useTextChangeHandler'
